Fun-dep on HasError and phantom on newtypes
The phantom parameter on @MonadThrow@ etc. is necessary so that deriving
via clauses can be written inline without having to specify the
exception type in a standalone deriving clause. Otherwise, deriving via
would fail due to an ambiguous exception type.

The functional-dependency is necessary for the same reason w.r.t. the
@Ctor@ combinator. Otherwise, instances cannot be inline derived via
@Ctor ctor@, and would require a standalone deriving clause to specify
the sum-type.

An alternative solution might be to add a phantom parameter to @Ctor@ to
make the sum-type explicit.