Capability-retaining `zoom` and `magnify`
This change allows callers of the `Capability.State.zoom` and `Capability.Reader.magnify` functions to pass along a set of capabilities from the current monadic context
to the focusing action. Retained capabilities are specified via a typelevel parameter of kind `[(* -> *) -> Constraint]`.

Implementation relies on `unsafeCoerce` to convert a dictionary for a wrapped monad instance to the original monad.

Fixes #46.