Don't re-export Capability.Constraint by default
Since we are advertising to import capability modules unqualified,
having a type family like `All` exported in these module would pollute
the global namespace. This is a bad idea.

I don't think any of the function in `Capability.Constraint` are
needed in the normal operations of Capability. But if we decide that
one is, we should re-export that one specifically.