Implement reflection for ad-hoc interpreters
Allows users to define ad-hoc interpretations of capabilities using the
functions `Capability.Reflection.interpret_` and
`Capability.Reflection.interpret` by providing a dictionary, i.e. a
record of functions, that defines an implementation of the capability.

The details are explained in the Haddock documentation of
`Capability.Reflection`. A reference to the top-level `Capability`
Haddock documentation is added as well.

Examples are defined in `examples/Reflection.hs`.