Add `censor` function to Capability.Writer
This is a useful function, present in the corresponding mtl library,
that can be defined in terms of the existing `pass` primitive.
However, doing so is nontrivial (I spent quite a while banging my head
against proxies, ScopedTypeVariables, AllowAmbiguousTypes, and the
like) so it seems it would make sense to provide it directly
from the library.  However, if there's some reason this function can't
or shouldn't be provided I'd be happy to be enlightened.